deleted vs erased with regards to a preservation lock
Remove "Content subject to the policy can't be modified or deleted during the retention period" as this implies that an end-user would be blocked from deleting the content. The content can be deleted but it will not be erased. The preservation lock will lock the policy itself but does not change experience for the end-user as this is still a container level retention policy. From the SEC17-a4 document, we phase it this way: 

Immutable, meaning that the record cannot be overwritten, ERASED, or altered during the required retention period.
In Exchange Online, when a retention policy is applied to a user's mailbox, all the user's content will be retained based on the criteria of the policy. In fact, if a user attempts to delete or modify an email, a copy of the email before the change is made will be preserved in a secure, hidden location in the user's mailbox. Retention policies can help ensure that an organization retains electronic communications, but those policies can be modified.

By placing a Preservation Lock on a retention policy, an organization ensures that the policy cannot be modified. In fact, after a Preservation Lock is applied to a retention policy, the following actions are restricted:

The retention period of the policy can only be increased, not shortened.
Users can be added to the policy, but no user can be removed.
The retention policy cannot be deleted by an administrator.